促销模块测试版

+++ anchongnet/app/Http/Controllers/admin/PromotionController.php
@@ -25,7 +25,7 @@ class PromotionController extends Controller
     /*
     *   到点促销开启
     */
-    public static function promotion()
+    public static function promotion($id)
     {
         //判断是否有促销缓存
         if (Cache::has('anchong_promotion_goods'))
@@ -35,10 +35,8 @@ class PromotionController extends Controller
         }
         //取到当前时间，为了避免时间不准时加十秒
         $nowtime=time()+10;
-        //查出当前促销时间段
-        $promotion_id=DB::table('anchong_promotion')->where('start_time','<',$nowtime)->where('end_time','>',$nowtime)->pluck('promotion_id');
         //查出促销时间段的商品
-        $goods_data=DB::table('anchong_promotion_goods')->where('promotion_id',$promotion_id[0])->select('gid','promotion_price')->orderBy('sort','DESC')->get();
+        $goods_data=DB::table('anchong_promotion_goods')->where('promotion_id',$id)->select('gid','promotion_price')->orderBy('sort','DESC')->get();
         //定义ORM模型
         $goods_specifications=new \App\Goods_specifications();
         //定义结果数组
@@ -70,6 +68,30 @@ class PromotionController extends Controller
         }
     }
 
+    /*
+    *   过点关闭上一波促销
+    */
+    public static function endpromotion($id)
+    {
+        //取到当前时间，为了避免时间不准时加十秒
+        $nowtime=time()+10;
+        //查出促销时间段的商品
+        $goods_data=DB::table('anchong_promotion_goods')->where('promotion_id',$id)->select('pg_id','gid')->orderBy('sort','DESC')->get();
+        //定义ORM模型
+        $goods_specifications=new \App\Goods_specifications();
+        //遍历数据并组合成新数据
+        foreach ($goods_data as $goodsinfo) {
+            //获得该货物信息的句柄
+            $goods_handle=$goods_specifications->find($goodsinfo->gid);
+            //修改其促销价
+            $goods_handle->promotion_price=0;
+            $save=$goods_handle->save();
+            //如果修改成功就促销该货品
+            if($save){
+                DB::table('anchong_promotion_goods')->where('pg_id',$goodsinfo->pg_id)->delete();
+            }
+        }
+    }
     /**
      * Show the form for creating a new resource.
      *
@@ -88,7 +110,26 @@ class PromotionController extends Controller
      */
     public function store(Request $request)
     {
-        //
+        $num=DB::table('anchong_promotion_goods')->where('promotion_id', $request->promotion_id)->count();
+        //判断是否达到该次促销数量的上限
+        if($num > 19){
+            return response()->json(['serverTime'=>time(),'ServerNo'=>19,'ResultData'=>['Message'=>'该时段的促销商品已达上限']]);
+        }
+        //将数据插入数据库
+        $result=DB::table('anchong_promotion_goods')->insert(
+            [
+                'promotion_id' => $request->promotion_id,
+                'gid' => $request->gid,
+                'promotion_price' => $request->promotion_price,
+                'sort' => $request->sort
+            ]
+        );
+        //判断是否插入成功
+        if($result){
+            return response()->json(['serverTime'=>time(),'ServerNo'=>0,'ResultData'=>['Message'=>'添加成功']]);
+        }else{
+            return response()->json(['serverTime'=>time(),'ServerNo'=>19,'ResultData'=>['Message'=>'添加失败']]);
+        }
     }
 
     /**
@@ -99,7 +140,15 @@ class PromotionController extends Controller
      */
     public function show($id)
     {
-        //
+        $nowtime=time();
+        $data=[];
+        //获取促销时间数据
+        $datas=DB::table('anchong_promotion')->select('promotion_id','start_time','end_time')->where('start_time','>',$nowtime)->take(5)->get();
+        //将unix时间戳转成日期
+        foreach ($datas as $value) {
+            $data[]=['promotion_id'=>$value->promotion_id,'start_time'=>date("Y-m-d H:i:s",$value->start_time),'end_time'=>date("Y-m-d H:i:s",$value->end_time)];
+        }
+        return $data;
     }
 
     /**
@@ -154,6 +203,10 @@ class PromotionController extends Controller
      */
     public function destroy($id)
     {
+        $num=DB::table('anchong_promotion_goods')->where('promotion_id',$id)->count();
+        if($num){
+            return "促销时间未结束，或该时段还有未下架的促销产品(请联系cto)！";
+        }
         //根据ID进行删除
         $result=DB::table('anchong_promotion')->where('promotion_id',$id)->delete();
         if($result){
